netdev-rte-offloads: Add default flow in vxlan vport table.
For packets that do not have a match in the vxlan vport table,
or in case the table is empty but still gets packets, need to
offload these flows and mark them with special mark.

In this case, the packets will arrive to ovs with that special
mark, and no table lookup will be needed.

Signed-off-by: Asaf Penso <asafp@mellanox.com>